refactor:#149 getLevelAndPercentage 함수 리팩토링  - 추후 레벨이 더 추가 및 삭제될 경우 유연하게 처리 가능하도록 함  - percent -> percentage가 더 정확한 표현이라 생각해 함수명도 수정  - 캐릭터 레벨 7까지 늘림
