+++ src/constants/character-constants.ts
@@ -19,12 +19,6 @@ type CharacterInfo = Record<string, CharacterDescription>;
 export const LEVEL_EXP = [0, 1500, 3000, 5000, 8000, 12000, 20000];
 
 export const CHARACTER_INFOMATIONS: CharacterInfo = {
-  // yellow: {
-  //   1: { name: '아기 병아리' },
-  //   2: { name: '안경 병아리' },
-  //   3: { name: '화난 병아리' },
-  //   4: { name: '미친 병아리' },
-  // },
   clay: {
     1: { name: '알 (클레이)' },
     2: { name: '학생 병아리' },
@@ -41,11 +35,6 @@ export const CHARACTER_INFOMATIONS: CharacterInfo = {
     5: { name: '정장 병아리' },
     6: { name: '불닭 병아리' },
   },
-  // 추후 확장 가능
-  // green: {
-  //   1: { name: '초록 병아리' },
-  //   ...
-  // },
 };
 
 export const CHARACTER_HISTORY_KEY = {
