+++ src/features/character/hooks/use-experience-up.ts
@@ -3,8 +3,6 @@ import { usePostExperienceMutation } from '@/features/character/hooks/use-post-e
 import { CHARACTER_HISTORY, CHARACTER_HISTORY_KEY } from '@/constants/character-constants';
 
 export const useExperienceUp = () => {
-  // const characterId = useCharacterStore((state) => state.characterId);
-
   const { mutate: expUpdateMutation } = usePostExperienceMutation();
 
   const handleExperienceUp = (history: keyof typeof CHARACTER_HISTORY_KEY) => {
